[mgsim-refactor] Avoid configuration defaults unless strictly necessary. Allow case-insensitive matches. Simplify the configuration code.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4366 e97e5017-a994-416e-809d-76780e9f78db

+++ sim/kernel.cpp
@@ -162,14 +162,14 @@ unsigned long long lcm(unsigned long long a, unsigned long long b)
     return (a * b / gcd(a,b));
 }
 
-Clock& Kernel::CreateClock(unsigned long long frequency)
+Clock& Kernel::CreateClock(unsigned long frequency)
 {
     // We only allow creating clocks before the simulation starts
     assert(m_cycle == 0);
 
     // Gotta be at least 1 MHz
     assert(frequency > 0);
-    frequency = std::max(1ULL, frequency);
+    frequency = std::max(1UL, frequency);
 
     /*
      Calculate new master frequency:
